backend/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
sqlalchemy
alembic
psycopg2-binary
pydantic-settings
python-dotenv
bcrypt
email-validator
python-multipart
